requirements.txt
asgiref==3.5.0
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
pytz==2021.3
certifi==2021.10.8
charset-normalizer==2.0.10
django-paypal==1.1.2
idna==3.3
Pillow==9.0.0
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stripe==2.65.0
urllib3==1.26.8
